Persist AI chat request_id and forward request metadata

+++ superset/migrations/versions/2026-06-24_14-45_064fcd8d1ef6_add_ai_chat_messages_table.py
@@ -35,7 +35,7 @@ def upgrade():
         "ai_chat_messages",
         sa.Column(
             "id",
-            sa.BigInteger(),
+            sa.Integer(),
             primary_key=True,
             autoincrement=True,
         ),
